Fix out-of memoery issue by adding a query-limit
Generating the admin-bar dropdown list was limited to 20 users already, but
the query to count the number of users still obtained a list of all users.
This commit limits the number of retrieved items and adds a "big website
theshold" constant that can be used to tune the shutoff threshold
for populating the adminbar dropdown with usernames.

Issue #22

Signed-off-by: Paul Konstantin Gerke <paulkgerke@craftware.info>